backend/requirements.txt
-i https://pypi.org/simple
altgraph==0.17.4
asgiref==3.7.2; python_version >= '3.7'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.1; python_full_version >= '3.7.0'
coverage==7.3.2; python_version >= '3.8'
django==4.2.6; python_version >= '3.8'
django-cors-headers==4.3.0; python_version >= '3.8'
djangorestframework==3.14.0; python_version >= '3.6'
drf-nested-routers==0.93.4; python_version >= '3.5'
google-api-core==2.12.0; python_version >= '3.7'
google-api-python-client==2.105.0; python_version >= '3.7'
google-auth==2.23.3; python_version >= '3.7'
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0; python_version >= '3.6'
googleapis-common-protos==1.61.0; python_version >= '3.7'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4; python_version >= '3.5'
importlib-metadata==6.8.0; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
protobuf==4.24.4; python_version >= '3.7'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyinstaller==6.1.0; python_version < '3.13' and python_version >= '3.8'
pyinstaller-hooks-contrib==2023.10; python_version >= '3.7'
pyparsing==3.1.1; python_full_version >= '3.6.8'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==68.2.2; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.8.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.7; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
